# core/region: use a separate tree to check for overlaps in the subtree