@rleungx 2024-05-17
BTW, how much extra memory will it cost?