@JmPotato 2024-05-17
> BTW, how much extra memory will it cost?

You can check the benchmark results in the PR description. It does not cause additional memory usage; on the contrary, it significantly reduces memory usage compared to previous versions.